SMILE-510 Moved sru-actions to new package.
