+++ src/test/java/no/unit/alma/GetAlmaSruRecordHandlerTest.java
@@ -1,5 +1,7 @@
 package no.unit.alma;
 
+import no.unit.alma.sru.AlmaSruConnection;
+import no.unit.alma.sru.GetAlmaSruRecordHandler;
 import org.junit.jupiter.api.Test;
 
 import javax.ws.rs.core.Response;
