+++ src/test/java/no/unit/alma/sru/cql/formatter/TestCqlFormatter.java
@@ -1,4 +1,4 @@
-package no.unit.cql.formatter;
+package no.unit.alma.sru.cql.formatter;
 
 import org.junit.jupiter.api.Test;
